requirements.txt
python-telegram-bot>=0
requests